python-mcp-server/pyproject.toml
[project]
name = "python-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "dotenv>=0.9.9",
    "fastapi>=0.116.2",
    "fastmcp>=0.4.1",
    "google>=3.0.0",
    "google-genai>=1.9.0",
    "google-generativeai>=0.5.4",
    "langchain>=0.3.23",
    "langchain-openai>=0.3.14",
    "mcp[cli]>=1.6.0",
    "ollama>=0.4.8",
    "pydantic-core>=2.33.1",
]
